Added a new comparison to the latest pass.
The new cuts dramatically improve the v3{2} and somewhat improve the v4{2},
although there are still problems there.  Unfortunately it would seem that
everything else suffers fairly significantly.  It's possible that very soon
we will have to drop the v4 entirely, which is fine.  Dropping the v3 is
harder so that means additional checks and adjustments will have to be made.